requirements.txt
selenium==4.21.0
undetected-chromedriver
schedule
python-dotenv==1.0.1

